# Conflicts:
#	src/Lykke.Service.PayInternal.Services/CalculationService.cs

+++ src/Lykke.Service.PayInternal.Contract/PaymentRequest/PaymentRequestDetailsMessage.cs
@@ -26,11 +26,11 @@ namespace Lykke.Service.PayInternal.Contract.PaymentRequest
         
         public DateTime DueDate { get; set; }
         
-        public double MarkupPercent { get; set; }
+        public decimal MarkupPercent { get; set; }
         
         public int MarkupPips { get; set; }
 
-        public double MarkupFixedFee { get; set; }
+        public decimal MarkupFixedFee { get; set; }
         
         public string WalletAddress { get; set; }
